backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
boto3
firecrawl-py
pydantic
requests
python-jose[cryptography]
